render: enable not importing image loaders and gift via build  flags
